Update refactoring (#6)
* feat : findById SwimmingPool Without Controller

* feat : findById and refactor mapping url

* feat : swimmingPool application layer Query Result Dto

* complete multi data into db. but it needs refactoring

+++ src/main/resources/h2/schema.sql
@@ -6,7 +6,7 @@ CREATE TABLE swimming_pool
     poolState     VARCHAR(50) NOT NULL,
     lotAddress    VARCHAR(100),
     streetAddress VARCHAR(100),
-    x             DOUBLE,
-    y             DOUBLE,
+    x             FLOAT,
+    y             FLOAT,
     PRIMARY KEY (poolId)
 );
